test(conformance): harden the differential conformance harness and corpus
Harness (tests/conformance_harness/):

- Cell identity and paths: replace the many-to-one workspace-name sanitizer with
  an injective slug encoder, so two distinct cells can no longer be handed the
  same workspace, and funnel every corpus path through one containment check that
  requires an absolute path, a regular file, residence under the canonicalized
  corpus root and the expected extension. A program and its expectation record
  must be siblings sharing a stem, and a symlink, directory or device node is
  refused rather than skipped.
- Expectation records: bound every read (whole-record, per-line, per-field and
  heredoc-length limits, re-checked after opening so the bound cannot be evaded
  between the size check and the read); require all eleven per-cell command
  placeholders; hold ref_command to a per-target driver as its first token, since
  the reference compiler has no target-selection flag and the driver binary is
  therefore the target; reject shell metacharacters in a template; require the
  mandatory static-linkage flag and reject any shared flag not verified as
  meaning the same thing to both compilers; restrict ub_audit_flags to the
  removable members of the audit gate; and refuse a record that no oracle could
  judge.
- Command rendering: render a template into an argument vector rather than a
  string, and quote each element only when a line is emitted for a human, so a
  path containing whitespace or shell grammar can neither be mis-split nor
  executed.
- Environment: read every variable through a checked accessor that fails on a
  non-UTF-8 value instead of defaulting silently, validate the whole variable
  catalogue before any process is spawned, and expose the run configuration as a
  single validated snapshot rather than a scattered set of accessors.
- Oracle discovery: vet each resolved tool for ownership, world-writability and
  search-path trust, confirm its target with -dumpmachine, require the compiler
  under test and the reference compiler to be different files, and bound every
  pre-flight probe with a deadline, an output cap and explicit reporting when a
  probe had to be terminated. An unavailable oracle is reported, never silently
  treated as a pass.
- Reports: escape tool banners, record fields and paths before they reach any
  report.

Corpus (tests/conformance/):

- Remove undefined and unspecified behaviour from five programs: a stride that
  formed a pointer two elements past the end of an array, pointer subtraction
  across the inner arrays of a multidimensional array, dependence on unspecified
  bitfield padding bits, and a pointer-to-integer conversion that is fatal on
  i686 with a 32-bit pointer.
- Strengthen the bitfield and pointer programs: print a byte image only for a
  struct whose every bit belongs to a named field, add a fully covered 32-bit
  layout whose fields cross every internal byte boundary, and add volatile
  runtime variants so each store must emit a real insert and each read a real
  extract at every optimization level.
- Condense the explanatory comments across all 73 programs and the include-path
  fixture into one register: no decorative banners, and each program restating
  its undefined-behaviour argument, its target-invariance derivation and its
  position with respect to the strict warning gate.
- Drop the placeholder file under tests/conformance/findings/.

Verified across the four supported targets: 292 strict-warning-gate compilations
with zero diagnostics, 73 sanitizer builds executed cleanly, and an 876-cell
compile-and-execute matrix (73 programs x 4 targets x 3 optimization levels) in
which every program's twelve cells produced byte-identical output.

+++ tests/conformance/10_declarations_and_types/007_alignof_alignas.c
@@ -1,8 +1,21 @@
-/* Area 10 / 007 -- _Alignof and _Alignas observed through alignment RESIDUES
- * and 0/1 relations. No address is ever printed. Only char, short, int and
- * float alignments are printed raw, because those four are identical on all
- * targets; double, long long, long and pointer alignments differ (4 on i686
- * versus 8 elsewhere) and appear only as bounded relations. */
+/* Alignment is observed through residues and 0/1 relations, and no address is
+ * ever printed.  Only the char, short, int and float alignments are printed raw,
+ * because those four are identical on all four targets; they are printed as
+ * values rather than asserted here, so a target that disagreed would show up as a
+ * divergence on that line.  The alignments of double, long long, long and
+ * pointers do differ (4 on i686 versus 8 elsewhere) and so appear only as bounded
+ * relations.
+ *
+ * WIDTH DISCIPLINE.  Each address is converted in two steps: first to unsigned
+ * long, which is exactly pointer width on all four targets (8 on x86-64, AArch64
+ * and RISC-V 64; 4 on i686), and only then widened to unsigned long long, which
+ * is 8 bytes everywhere and so is always wide enough to hold the residue
+ * arithmetic below.  Casting a pointer straight to unsigned long long would cross
+ * a width boundary on i686 and is diagnosed as -Wpointer-to-int-cast, which the
+ * mandatory audit gate escalates to an error; the two-step form never crosses
+ * one.  This matches the carrier discipline used by
+ * 05_pointers/007_casts_roundtrip.c.  uintptr_t is not used because it would
+ * require stdint.h and no header may be included. */
 
 int printf(const char *, ...);
 
@@ -15,10 +28,14 @@ int main(void)
 {
     _Alignas(8) unsigned char local8[8];
     struct padded p;
-    unsigned long long a_buf = (unsigned long long)(const void *)buf16;
-    unsigned long long a_wide = (unsigned long long)(const void *)&wide_aligned;
-    unsigned long long a_local = (unsigned long long)(const void *)local8;
-    unsigned long long a_member = (unsigned long long)(const void *)&p.aligned_member;
+    unsigned long long a_buf =
+        (unsigned long long)(unsigned long)(const void *)buf16;
+    unsigned long long a_wide =
+        (unsigned long long)(unsigned long)(const void *)&wide_aligned;
+    unsigned long long a_local =
+        (unsigned long long)(unsigned long)(const void *)local8;
+    unsigned long long a_member =
+        (unsigned long long)(unsigned long)(const void *)&p.aligned_member;
 
     local8[0] = 1;
     p.c = 2;
